Make buffer averaging more efficient.
Instead of iterating over the buffer every time we want the average,
which will presumably be done in a tight loop, we just maintain a sum of
the elements in the buffer as they are inserted. This makes averaging a
O(1) operation instead of O(n), and insertion remains O(1).